@@ -25,98 +26,126 @@
   "author": "Sandrino Di Mattia",
   "license": "MIT",
   "dependencies": {
-    "async": "~1.5.2",
-    "auth0": "^2.1.0",
-    "auth0-oauth2-express": "^1.2.0",
-    "aws-sdk": "^2.2.47",
-    "axios": "^0.9.1",
-    "babel": "^6.5.2",
-    "babel-core": "^6.7.4",
-    "babel-plugin-transform-runtime": "^6.6.0",
-    "babel-polyfill": "^6.7.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-es2015-loose": "^7.0.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.7.2",
-    "babel-runtime": "^6.6.1",
-    "bluebird": "^3.3.4",
-    "body-parser": "^1.15.0",
-    "ejs": "^2.4.1",
-    "express": "^4.13.4",
-    "express-jwt": "^3.3.0",
+    "async": "^1.5.2",
+    "auth0": "^2.4.0",
+    "auth0-extension-hapi-tools": "1.0.0",
+    "auth0-extension-tools": "^1.0.0",
+    "auth0-extension-ui": "git+https://github.com/auth0-extensions/auth0-extension-ui",
+    "auth0-extension-ui-redux": "git+https://github.com/auth0-extensions/auth0-extension-ui-redux",
+    "aws-sdk": "^2.3.17",
+    "axios": "^0.15.0",
+    "blipp": "^2.3.0",
+    "bluebird": "^3.4.0",
+    "body-parser": "^1.15.1",
+    "boom": "3.2.2",
+    "ejs": "^2.4.2",
+    "expect": "^1.20.2",
     "glob": "^7.0.3",
+    "good": "7.0.1",
+    "good-console": "6.1.2",
+    "hapi": "13.5.0",
+    "hapi-auth-jwt2": "7.0.1",
+    "hapi-swagger": "7.4.0",
+    "inert": "^4.0.1",
+    "joi": "9.0.4",
     "json-loader": "^0.5.4",
-    "lodash": "^3.9.3",
-    "lowdb": "^0.12.5",
-    "lru-memoizer": "^1.1.0",
-    "moment": "^2.12.0",
-    "mongo-getdb": "^2.0.0",
-    "mongodb-uri": "^0.9.7",
-    "morgan": "^1.7.0",
+    "jsonwebtoken": "^7.0.0",
+    "jwks-rsa": "^1.1.1",
+    "lodash": "^3.10.1",
+    "lru-memoizer": "1.10.0",
+    "moment": "^2.13.0",
+    "ms": "^0.7.1",
     "nconf": "^0.8.4",
     "node-uuid": "^1.4.7",
-    "request": "^2.69.0",
+    "relish": "^0.2.3",
+    "request": "^2.72.0",
     "semver": "^5.1.0",
-    "validate.js": "^0.9.0",
-    "webtask-tools": "^1.3.0",
+    "superagent": "^2.3.0",
+    "webtask-tools": "^3.0.0",
     "winston": "^2.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.3.4",
-    "babel-eslint": "^6.0.0",
-    "babel-loader": "^6.2.4",
+    "auth0-styleguide": "auth0/styleguide#4.6.5",
+    "autoprefixer": "^6.5.0",
+    "babel": "^6.5.2",
+    "babel-core": "^6.17.0",
+    "babel-eslint": "7.1.1",
+    "babel-loader": "6.2.8",
+    "babel-plugin-transform-runtime": "^6.15.0",
+    "babel-polyfill": "^6.16.0",
+    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-es2015-loose": "^8.0.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-stage-0": "^6.16.0",
+    "babel-register": "^6.16.3",
+    "babel-runtime": "^6.9.2",
     "chai": "^3.5.0",
-    "classnames": "^2.2.3",
-    "css-loader": "^0.23.1",
-    "eslint": "^2.4.0",
-    "eslint-config-airbnb": "^6.2.0",
-    "eslint-plugin-babel": "^3.1.0",
-    "eslint-plugin-react": "^4.2.3",
+    "classnames": "^2.2.5",
+    "cross-env": "^3.1.2",
+    "css-loader": "0.26.0",
+    "eslint": "^3.10.1",
+    "eslint-config-auth0": "^6.0.1",
+    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-import": "^1.16.0",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-react": "^6.7.1",
     "exports-loader": "^0.6.3",
     "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
-    "history": "^2.0.1",
-    "immutable": "^3.7.6",
+    "file-loader": "^0.9.0",
+    "gulp": "^3.9.1",
+    "gulp-cli": "^1.2.2",
+    "gulp-nodemon": "^2.2.1",
+    "gulp-util": "^3.0.7",
+    "history": "^2.1.2",
+    "immutable": "^3.8.1",
     "imports-loader": "^0.6.5",
     "json-loader": "^0.5.4",
     "jwt-decode": "^2.0.1",
-    "mocha": "^2.4.5",
-    "nodemon": "^1.9.1",
+    "mocha": "^3.1.1",
+    "ngrok": "^2.2.3",
+    "nock": "^9.0.2",
+    "nodemon": "^1.11.0",
+    "nyc": "10.0.0",
+    "open": "0.0.5",
     "postcss-focus": "^1.0.0",
-    "postcss-loader": "^0.8.2",
-    "postcss-reporter": "^1.3.3",
-    "postcss-simple-vars": "^1.2.0",
+    "postcss-loader": "^1.1.1",
+    "postcss-reporter": "^2.0.0",
+    "poststylus": "^0.2.3",
+    "qs": "^6.2.1",
     "raw-loader": "^0.5.1",
-    "react": "^0.14.7",
-    "react-bootstrap": "^0.28.4",
-    "react-dom": "^0.14.7",
-    "react-loader-advanced": "^1.1.2",
+    "react": "^15.3.2",
+    "react-bootstrap": "^0.30.5",
+    "react-dom": "^15.3.2",
+    "react-loader-advanced": "^1.4.0",
     "react-pure-render": "^1.0.2",
-    "react-redux": "^4.4.1",
-    "react-router": "^2.0.1",
-    "react-router-redux": "^4.0.0",
+    "react-redux": "^4.4.5",
+    "react-router": "^2.8.1",
+    "react-router-redux": "^4.0.6",
+    "react-select": "^1.0.0-rc.1",
     "react-transform-hmr": "^1.0.4",
-    "redux": "^3.3.1",
-    "redux-devtools": "^3.1.1",
-    "redux-devtools-dock-monitor": "^1.1.0",
-    "redux-devtools-log-monitor": "^1.0.5",
-    "redux-form": "^4.2.2",
-    "redux-logger": "^2.6.1",
-    "redux-promise-middleware": "^2.3.3",
+    "redux": "^3.6.0",
+    "redux-devtools": "^3.3.1",
+    "redux-devtools-dock-monitor": "^1.1.1",
+    "redux-devtools-log-monitor": "^1.0.11",
+    "redux-form": "^6.0.5",
+    "redux-logger": "^2.7.0",
+    "redux-promise-middleware": "^4.1.0",
     "redux-simple-router": "^2.0.4",
-    "redux-thunk": "^2.0.1",
+    "redux-static": "^1.0.0",
+    "redux-thunk": "^2.1.0",
     "request-promise": "^2.0.1",
     "rimraf": "^2.5.2",
     "style-loader": "^0.13.1",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^2.3.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.12.14",
-    "webpack-dev-middleware": "^1.6.0",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-hot-middleware": "^2.10.0",
-    "webpack-sources": "^0.1.1",
+    "webpack": "^1.13.2",
+    "webpack-dev-middleware": "^1.8.4",
+    "webpack-dev-server": "^1.16.2",
+    "webpack-hot-middleware": "^2.13.2",
+    "webpack-sources": "^0.1.3",
     "webpack-stats-plugin": "^0.1.1",
-    "webtask-bundle": "^1.3.2"
+    "webtask-bundle": "^2.1.1"
   }
 }
