@@ -1,95 +1,86 @@
-import path from 'path';
-import jwtDecode from 'jwt-decode';
-import Express from 'express';
-import morgan from 'morgan';
-import nconf from 'nconf';
-import bodyParser from 'body-parser';
-import validator from 'validate.js';
-import auth0 from 'auth0-oauth2-express';
+import Hapi from 'hapi';
+import Good from 'good';
+import Inert from 'inert';
+import Relish from 'relish';
+import Blipp from 'blipp';
+import jwt from 'hapi-auth-jwt2';
+import GoodConsole from 'good-console';
+import HapiSwagger from 'hapi-swagger';
 
-import Database from './lib/storage/database';
-import { S3Provider } from './lib/storage/providers';
-import { init as initDb, getDb } from './lib/storage/getdb';
-import api from './routes/api';
-import meta from './routes/meta';
-import htmlRoute from './routes/html';
+import config from './lib/config';
 import logger from './lib/logger';
-import ensureRule from './lib/ensureRule';
-import * as middlewares from './lib/middlewares';
+import plugins from './plugins';
 
-module.exports = (options = { }) => {
-  // Configure validator.
-  validator.options = { fullMessages: false };
-  validator.validators.presence.options = {
-    message: (value, attribute) => `The ${attribute} is required.`
+export default (cb) => {
+  const goodPlugin = {
+    register: Good,
+    options: {
+      ops: {
+        interval: 30000
+      },
+      reporters: {
+        console: [
+
+        ]
+      }
+    }
   };
 
-  // Initialize database.
-  initDb(new Database({
-    provider: options.storageProvider || new S3Provider({
-      path: 'auth0-authz.json',
-      bucket: nconf.get('AWS_S3_BUCKET'),
-      keyId: nconf.get('AWS_ACCESS_KEY_ID'),
-      keySecret: nconf.get('AWS_SECRET_ACCESS_KEY')
-    })
-  }));
+  const hapiSwaggerPlugin = {
+    register: HapiSwagger,
+    options: {
+      documentationPage: false,
+      swaggerUI: false
+    }
+  };
 
-  // Initialize the app.
-  const app = new Express();
-  app.use(morgan(':method :url :status :response-time ms - :res[content-length]', {
-    stream: logger.stream
-  }));
-  app.use(bodyParser.json());
-  app.use(bodyParser.urlencoded({ extended: false }));
-
-  // Configure routes.
-  app.use('/api', api());
-  app.use('/app', Express.static(path.join(__dirname, '../dist')));
-  app.use('/meta', meta());
-
-  // Use OAuth2 authorization.
-  if (nconf.get('USE_OAUTH2')) {
-    // Helper to store the user's token in storage.
-    const onUserAuthenticated = (req, res, accessToken, next) => {
-      const decodedToken = jwtDecode(accessToken);
-      getDb().setToken(decodedToken.sub, { accessToken })
-        .then(() => {
-          // Create the rule.
-          ensureRule(accessToken);
-
-          // Continue.
-          next();
-        })
-        .catch(next);
-    };
-
-    // Authenticate non-admins.
-    // app.use(auth0({
-    //   scopes: nconf.get('AUTH0_SCOPES'),
-    //   authenticatedCallback: onUserAuthenticated,
-    //   clientId: nconf.get('AUTH0_CLIENT_ID'),
-    //   rootTenantAuthority: `https://${nconf.get('AUTH0_DOMAIN')}`,
-    //   apiToken: {
-    //     secret: nconf.get('AUTHORIZE_API_KEY')
-    //   }
-    // }));
-
-    // Authenticate admins.
-    app.use('/admins', auth0({
-      scopes: nconf.get('AUTH0_SCOPES'),
-      authenticatedCallback: onUserAuthenticated,
-      clientName: 'Auth0 Authorization Dashboard Extension',
-      apiToken: {
-        secret: nconf.get('AUTHORIZE_API_KEY')
-      },
-      audience: (req) => `https://${req.webtaskContext.data.AUTH0_DOMAIN}/api/v2/`
-    }));
+  if (process.env.NODE_ENV !== 'test') {
+    goodPlugin.options.reporters.console.push(
+      new GoodConsole({ color: !!config('LOG_COLOR') })
+    );
+    goodPlugin.options.reporters.console.push('stdout');
   }
 
-  // Fallback to rendering HTML.
-  app.get('*', htmlRoute());
+  const relishPlugin = Relish({ });
 
-  // Generic error handler.
-  app.use(middlewares.errorHandler);
-  return app;
+  const server = new Hapi.Server();
+  server.connection({
+    host: 'localhost',
+    port: config('PORT'),
+    routes: {
+      cors: true,
+      validate: {
+        failAction: relishPlugin.failAction
+      }
+    }
+  });
+  server.register([ goodPlugin, Inert, Blipp, jwt, hapiSwaggerPlugin, ...plugins ], (err) => {
+    if (err) {
+      return cb(err, null);
+    }
+
+    // Use the server logger.
+    logger.debug = (...args) => {
+      server.log([ 'debug' ], args.join(' '));
+    };
+    logger.info = (...args) => {
+      server.log([ 'info' ], args.join(' '));
+    };
+    logger.error = (...args) => {
+      server.log([ 'error' ], args.join(' '));
+    };
+
+    return cb(null, server);
+  });
+
+  server.ext('onPreResponse', (request, reply) => {
+    if (request.response && request.response.isBoom && request.response.output) {
+      server.log([ 'error' ], `Request: ${request.method.toUpperCase()} ${request.url.path}`);
+      server.log([ 'error' ], `Response: ${JSON.stringify(request.response, null, 2)}`);
+    }
+
+    return reply.continue();
+  });
+
+  return server;
 };
