@@ -12,7 +12,20 @@ export function fetchUserGroups(userId) {
     },
     payload: {
       promise: axios.get(`/api/users/${userId}/groups`, {
-        timeout: 5000,
+        responseType: 'json'
+      })
+    }
+  };
+}
+
+export function fetchUserNestedGroups(userId) {
+  return {
+    type: constants.FETCH_USER_NESTED_GROUPS,
+    meta: {
+      userId
+    },
+    payload: {
+      promise: axios.get(`/api/users/${userId}/groups/calculate`, {
         responseType: 'json'
       })
     }
