@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Error, Confirm, LoadingPanel } from '../Dashboard';
+import { Error, Confirm, LoadingPanel } from 'auth0-extension-ui';
 
 class GroupDeleteDialog extends Component {
   constructor() {
@@ -11,11 +11,11 @@ class GroupDeleteDialog extends Component {
   shouldComponentUpdate(nextProps) {
     return nextProps.group !== this.props.group;
   }
-  
+
   confirm() {
     this.props.onConfirm(this.props.group.get('record').toJS());
   }
-  
+
   clear() {
     this.props.onCancel();
   }
@@ -25,14 +25,18 @@ class GroupDeleteDialog extends Component {
     const title = `Delete Group: ${group.record.name}`;
 
     return (
-      <Confirm title={ title } show={ group.requesting && group.isDelete } loading={ group.loading } 
-        onCancel={ this.clear } onConfirm={ this.confirm }>
-        <LoadingPanel show={ group.loading } spinnerStyle={{ height: '40px', width: '40px' }}
-            animationStyle={{ paddingTop: '5px', paddingBottom: '5px' }}>
-          <Error message={ group.error } />
-          <p>
-            Do you really want to delete "<strong>{ group.record.name }</strong>"? 
-            This means all group memberships will also be removed.
+      <Confirm
+        title={title} show={group.requesting && group.isDelete} loading={group.loading}
+        onCancel={this.clear} onConfirm={this.confirm} confirmMessage="Delete"
+      >
+        <LoadingPanel
+          show={group.loading} spinnerStyle={{ height: '40px', width: '40px' }}
+          animationStyle={{ paddingTop: '5px', paddingBottom: '5px' }}
+        >
+          <Error message={group.error} />
+          <p className="text-center">
+            Do you really want to delete "<strong>{ group.record.name }</strong>"? <br />
+            This means all group mappings and memberships will also be removed.
           </p>
         </LoadingPanel>
       </Confirm>
