@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Modal } from 'react-bootstrap';
 
-import { Error } from '../Dashboard';
+import { Error } from 'auth0-extension-ui';
 import GroupForm from './GroupForm';
 
 class GroupDialog extends Component {
@@ -11,18 +11,26 @@ class GroupDialog extends Component {
 
   render() {
     const group = this.props.group.toJS();
-    const title = group.isNew ? 'New Group' : `Edit Group: ${group.record.name}`;
+    const title = group.isNew ? 'Create Group' : `Edit Group: ${group.record.name}`;
     const isVisible = group.isEdit || group.isNew;
 
-    return <Modal show={isVisible} onHide={this.props.onClose}>
-      <Modal.Header closeButton={!group.loading}>
-        <Modal.Title>{title}</Modal.Title>
-      </Modal.Header>
-      <GroupForm loading={group.loading} initialValues={group.record} validationErrors={group.validationErrors}
-          onClose={this.props.onClose} onSubmit={(perm) => this.props.onSave(perm)}>
-        <Error message={group.error} />
-      </GroupForm>
-    </Modal>;
+    return (
+      <Modal show={isVisible} className="modal-overflow-visible" onHide={this.props.onClose}>
+        <Modal.Header closeButton={!group.loading} className="has-border">
+          <Modal.Title>{title}</Modal.Title>
+        </Modal.Header>
+        <GroupForm
+          loading={group.loading}
+          initialValues={group.record}
+          validationErrors={group.validationErrors}
+          onClose={this.props.onClose}
+          onSubmit={(perm) => this.props.onSave(perm)}
+          isNew={group.isNew}
+        >
+          <Error message={group.error} />
+        </GroupForm>
+      </Modal>
+    );
   }
 }
 
