@@ -1,41 +1,49 @@
 import React, { Component } from 'react';
-import { reduxForm } from 'redux-form';
-import { Button, ButtonToolbar, Modal } from 'react-bootstrap';
+import { Field } from 'redux-form';
+import { Button, Modal } from 'react-bootstrap';
+import { InputText, LoadingPanel } from 'auth0-extension-ui';
 
-import { InputText, LoadingPanel } from '../Dashboard';
+import createForm from '../../utils/createForm';
 
-class GroupForm extends Component {
+export default createForm('group', class GroupForm extends Component {
+  static propTypes = {
+    validationErrors: React.PropTypes.object,
+    loading: React.PropTypes.bool.isRequired,
+    submitting: React.PropTypes.bool,
+    handleSubmit: React.PropTypes.func.isRequired,
+    onClose: React.PropTypes.func.isRequired,
+    children: React.PropTypes.node
+  };
+
   render() {
-    const { fields: { name, description }, handleSubmit, loading, submitting, validationErrors } = this.props;
-
-    return <div>
+    const { handleSubmit, loading, submitting, validationErrors, isNew } = this.props;
+    return (<div>
       <Modal.Body>
-        {this.props.children}
         <LoadingPanel show={loading}>
-          <InputText field={name} fieldName="name" label="Name"
-            validationErrors={validationErrors} />
-          <InputText field={description} fieldName="description" label="Description"
-            validationErrors={validationErrors} />
+          { isNew &&
+            <p className="modal-description">
+              Give your group a name. After creating the group you'll be able to assing users and roles to it.
+            </p>
+          }
+          {this.props.children}
+          <Field
+            name="name" component={InputText} label="Name"
+            validationErrors={validationErrors}
+          />
+          <Field
+            name="description" component={InputText}
+            label="Description" validationErrors={validationErrors}
+          />
         </LoadingPanel>
       </Modal.Body>
       <Modal.Footer>
-        <Button bsSize="small" disabled={ loading || submitting } onClick={this.props.onClose}>
+        <Button bsSize="large" bsStyle="transparent" disabled={loading || submitting} onClick={this.props.onClose}>
           Cancel
         </Button>
-        <Button bsStyle="primary" bsSize="small" disabled={ loading || submitting } onClick={handleSubmit}>
-          Save
+        <Button bsSize="large" bsStyle="primary" disabled={loading || submitting} onClick={handleSubmit}>
+          { isNew ? 'Create' : 'Save' }
         </Button>
       </Modal.Footer>
-    </div>;
+    </div>);
   }
-}
-
-GroupForm.propTypes = {
-  validationErrors: React.PropTypes.object,
-  loading: React.PropTypes.bool.isRequired,
-  submitting: React.PropTypes.bool,
-  handleSubmit: React.PropTypes.func.isRequired,
-  onClose: React.PropTypes.func.isRequired
-};
-
-export default reduxForm({ form: 'group', fields: [ 'name', 'description' ] })(GroupForm);
+});
