@@ -1,12 +1,24 @@
 import React, { Component } from 'react';
-import { reduxForm } from 'redux-form';
-import { Button, ButtonToolbar, Modal } from 'react-bootstrap';
+import { Field } from 'redux-form';
+import { Button, Modal } from 'react-bootstrap';
 
-import { InputText, InputCombo, LoadingPanel } from '../Dashboard';
+import { InputText, InputCombo, LoadingPanel } from 'auth0-extension-ui';
+import createForm from '../../utils/createForm';
 
-class GroupMappingForm extends Component {
+export default createForm('groupMapping', class GroupMappingForm extends Component {
+  propTypes = {
+    children: React.PropTypes.object,
+    fields: React.PropTypes.object,
+    connections: React.PropTypes.array,
+    validationErrors: React.PropTypes.object,
+    loading: React.PropTypes.bool.isRequired,
+    submitting: React.PropTypes.bool,
+    handleSubmit: React.PropTypes.func.isRequired,
+    onClose: React.PropTypes.func.isRequired
+  };
+
   render() {
-    const { fields: { connectionName, groupName }, handleSubmit, loading, submitting, validationErrors } = this.props;
+    const { handleSubmit, loading, submitting, validationErrors } = this.props;
 
     const connections = this.props.connections.map(connection => ({
       value: connection.name,
@@ -17,33 +29,33 @@ class GroupMappingForm extends Component {
       <div>
         <Modal.Body>
           {this.props.children}
-          <LoadingPanel show={ loading } spinnerStyle={{ height: '16px', width: '16px' }} animationStyle={{ paddingTop: '0px', paddingBottom: '0px', marginTop: '0px', marginBottom: '10px' }}>
-            <InputCombo options={connections} field={connectionName} fieldName="connectionName" label="Connection" validationErrors={validationErrors} />
-            <InputText field={groupName} fieldName="groupName" label="Incoming Group Name" validationErrors={validationErrors} />
+          <LoadingPanel show={loading} spinnerStyle={{ height: '16px', width: '16px' }} animationStyle={{ paddingTop: '0px', paddingBottom: '0px', marginTop: '0px', marginBottom: '10px' }}>
+            <p className="modal-description">
+              When users log in with the selected connection and belong to the selected group then they will also become an implicit member of this groupi.
+            </p>
+
+            <Field
+              name="connectionName" component={InputCombo}
+              options={connections} label="Connection"
+              validationErrors={validationErrors}
+            />
+
+            <Field
+              name="groupName" component={InputText} label="Incoming Group Name"
+              validationErrors={validationErrors}
+            />
+
           </LoadingPanel>
         </Modal.Body>
         <Modal.Footer>
-          <Button bsSize="small" disabled={ loading || submitting } onClick={this.props.onClose}>
+          <Button disabled={loading || submitting} onClick={this.props.onClose}>
             Cancel
           </Button>
-          <Button bsStyle="primary" bsSize="small" disabled={ loading || submitting } onClick={handleSubmit}>
-            Save
+          <Button bsStyle="primary" disabled={loading || submitting} onClick={handleSubmit}>
+            Create
           </Button>
         </Modal.Footer>
       </div>
     );
   }
-}
-
-GroupMappingForm.propTypes = {
-  children: React.PropTypes.object,
-  fields: React.PropTypes.object,
-  connections: React.PropTypes.array,
-  validationErrors: React.PropTypes.object,
-  loading: React.PropTypes.bool.isRequired,
-  submitting: React.PropTypes.bool,
-  handleSubmit: React.PropTypes.func.isRequired,
-  onClose: React.PropTypes.func.isRequired
-};
-
-export default reduxForm({ form: 'groupMapping', fields: [ 'connectionName', 'groupName' ] })(GroupMappingForm);
+});
