@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react';
-import { TableAction } from '../Dashboard';
+import { TableAction } from 'auth0-extension-ui';
 
 class GroupMappingRemoveAction extends Component {
   constructor() {
@@ -14,7 +14,8 @@ class GroupMappingRemoveAction extends Component {
   render() {
     const { index, loading } = this.props;
     return (
-      <TableAction id={`remove-groupMapping-${index}`} type="success" title="Remove" icon="263"
+      <TableAction
+        id={`remove-groupMapping-${index}`} title="Remove" icon="471"
         onClick={this.onRemove} disabled={loading || false}
       />
     );
