@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Error, Confirm, LoadingPanel } from '../Dashboard';
+import { Error, Confirm, LoadingPanel } from 'auth0-extension-ui';
 
 class GroupMappingRemoveDialog extends Component {
   constructor() {
@@ -26,9 +26,9 @@ class GroupMappingRemoveDialog extends Component {
 
     const title = `Remove mapping from ${group.record.name}`;
     return (
-      <Confirm title={ title } show={ requesting && isDelete } loading={ loading } onCancel={ this.clear } onConfirm={ this.confirm }>
-        <LoadingPanel show={ loading } spinnerStyle={{ height: '40px', width: '40px' }} animationStyle={{ paddingTop: '5px', paddingBottom: '5px' }}>
-          <Error message={ error } />
+      <Confirm title={title} show={requesting && isDelete} loading={loading} onCancel={this.clear} onConfirm={this.confirm}>
+        <LoadingPanel show={loading} spinnerStyle={{ height: '40px', width: '40px' }} animationStyle={{ paddingTop: '5px', paddingBottom: '5px' }}>
+          <Error message={error} />
           <p>Do you really want to remove the following mapping from <strong>{ group.record.name }</strong>?</p>
           <p><strong>{ record.connectionName }</strong> / <strong>{ record.groupName }</strong></p>
         </LoadingPanel>
