@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Button, ButtonToolbar } from 'react-bootstrap';
-import { Error, LoadingPanel, Table, TableCell, TableBody, TableIconCell, TableTextCell, TableHeader, TableColumn, TableRow } from '../Dashboard';
+import { Error, LoadingPanel, Table, TableCell, TableBody, TableIconCell, TableTextCell, TableHeader, TableColumn, TableRow } from 'auth0-extension-ui';
 
 import { GroupMappingRemoveAction } from './';
 
@@ -21,15 +21,16 @@ class GroupMappings extends Component {
             </div>
           </div>
           <div className="row">
-            <div className="col-xs-10">
-              <span className="pull-left">Mappings allow you to define which existing group memberships (eg: Active Directory) should be translated to this group. For example: &quot;<i>If <strong>John</strong> from <strong>fabrikam-adfs</strong> belongs to <strong>Fabrikam HR</strong> then he should also be member of this group.&quot;</i></span>
+            <div className="col-xs-8">
+              <p>
+                Mappings allow you to define which existing group memberships (eg: Active Directory) should be translated to this group.
+                For example: &quot;<i>If <strong>John</strong> from <strong>fabrikam-adfs</strong> belongs to <strong>Fabrikam HR</strong> then he should also be member of this group.&quot;</i>
+              </p>
             </div>
-            <div className="col-xs-2">
-              <ButtonToolbar className="pull-right">
-                <Button bsStyle="primary" bsSize="xsmall" onClick={this.props.createMapping} disabled={loading}>
-                  <i className="icon icon-budicon-337"></i> Create
-                </Button>
-              </ButtonToolbar>
+            <div className="col-xs-4">
+              <Button className="pull-right" bsStyle="success" onClick={this.props.createMapping} disabled={loading}>
+                <i className="icon icon-budicon-473" /> Create group mapping
+              </Button>
             </div>
           </div>
           <div className="row">
@@ -42,17 +43,17 @@ class GroupMappings extends Component {
                   <TableColumn width="10%" />
                 </TableHeader>
                 <TableBody>
-                {records.map((mapping, index) => (
-                  <TableRow key={index}>
-                    <TableIconCell color="green" icon="573" />
-                    <TableTextCell>{mapping.connectionName}</TableTextCell>
-                    <TableTextCell>{mapping.groupName}</TableTextCell>
-                    <TableCell>
-                      <ButtonToolbar style={{ marginBottom: '0px' }}>
-                        <GroupMappingRemoveAction index={index} groupMapping={mapping} loading={loading} onRemove={this.props.removeMapping} />
-                      </ButtonToolbar>
-                    </TableCell>
-                  </TableRow>
+                  {records.map((mapping, index) => (
+                    <TableRow key={index}>
+                      <TableIconCell color="green" icon="573" />
+                      <TableTextCell>{mapping.connectionName}</TableTextCell>
+                      <TableTextCell>{mapping.groupName}</TableTextCell>
+                      <TableCell>
+                        <ButtonToolbar style={{ marginBottom: '0px' }}>
+                          <GroupMappingRemoveAction index={index} groupMapping={mapping} loading={loading} onRemove={this.props.removeMapping} />
+                        </ButtonToolbar>
+                      </TableCell>
+                    </TableRow>
                 ))}
                 </TableBody>
               </Table>
