@@ -1,23 +1,25 @@
 import React, { Component } from 'react';
 import { Button, ButtonToolbar, Modal } from 'react-bootstrap';
 
-import { Error, Json, LoadingPanel } from '../Dashboard';
+import { Error, Json, LoadingPanel } from 'auth0-extension-ui';
 
 class LogDialog extends Component {
   render() {
     const { logId, error, loading, onClose } = this.props;
     if (logId === null) {
-      return <div></div>;
+      return <div />;
     }
 
     const log = this.props.log.toJS();
-    return <Modal show={logId !== null} onHide={onClose}>
+    return (<Modal show={logId !== null} onHide={onClose}>
       <Modal.Header closeButton={!loading}>
         <Modal.Title>Log - <span>{log.type || 'Log Record'}</span></Modal.Title>
       </Modal.Header>
       <Modal.Body>
-        <LoadingPanel show={loading} spinnerStyle={{ height: '16px', width: '16px' }}
-            animationStyle={{ paddingTop: '0px', paddingBottom: '0px', marginTop: '0px', marginBottom: '10px' }}>
+        <LoadingPanel
+          show={loading} spinnerStyle={{ height: '16px', width: '16px' }}
+          animationStyle={{ paddingTop: '0px', paddingBottom: '0px', marginTop: '0px', marginBottom: '10px' }}
+        >
           <Error message={error}>
             <Json jsonObject={log} />
           </Error>
@@ -26,11 +28,11 @@ class LogDialog extends Component {
       <Modal.Footer>
         <ButtonToolbar>
           <Button bsSize="small" disabled={loading} onClick={onClose}>
-            Cancel
+            <i className="icon icon-budicon-501" /> Cancel
           </Button>
         </ButtonToolbar>
       </Modal.Footer>
-    </Modal>;
+    </Modal>);
   }
 }
 
