@@ -1,37 +1,40 @@
-import React, { Component } from 'react';
+import React, { Component, PropTypes } from 'react';
 import { Modal } from 'react-bootstrap';
 
-import { Error } from '../Dashboard';
+import { Error } from 'auth0-extension-ui';
 import PermissionForm from './PermissionForm';
 
-class PermissionDialog extends Component {
+export default class PermissionDialog extends Component {
+  static propTypes = {
+    onSave: PropTypes.func.isRequired,
+    onClose: PropTypes.func.isRequired,
+    permission: PropTypes.object.isRequired,
+    applications: PropTypes.object.isRequired
+  }
+
   shouldComponentUpdate(nextProps) {
     return nextProps.permission !== this.props.permission || nextProps.applications !== this.props.applications;
   }
 
   render() {
-    const permission = this.props.permission.toJS();
     const applications = this.props.applications.toJS();
-
-    const title = permission.isNew ? 'New Permission' : `Edit Permission: ${permission.record.name}`;
+    const permission = this.props.permission.toJS();
+    const title = permission.isNew ? 'Create Permission' : `Edit Permission: ${permission.record.name}`;
     const isVisible = permission.isEdit || permission.isNew;
 
-    return <Modal show={isVisible} onHide={this.props.onClose}>
-      <Modal.Header closeButton={!permission.loading}>
-        <Modal.Title>{title}</Modal.Title>
-      </Modal.Header>
-      <PermissionForm loading={permission.loading} applications={applications} initialValues={permission.record} validationErrors={permission.validationErrors}
-          onClose={this.props.onClose} onSubmit={(perm) => this.props.onSave(perm)}>
+    return (
+      <Modal show={isVisible} onHide={this.props.onClose}>
+        <Modal.Header closeButton={!permission.loading} className="has-border">
+          <Modal.Title>{title}</Modal.Title>
+        </Modal.Header>
+        <PermissionForm
+          applications={applications.records} loading={permission.loading} initialValues={permission.record} validationErrors={permission.validationErrors}
+          onClose={this.props.onClose} onSubmit={this.props.onSave}
+          isNew={permission.isNew}
+        >
           <Error message={permission.error} />
-      </PermissionForm>
-    </Modal>;
+        </PermissionForm>
+      </Modal>
+    );
   }
 }
-
-PermissionDialog.propTypes = {
-  onSave: React.PropTypes.func.isRequired,
-  onClose: React.PropTypes.func.isRequired,
-  permission: React.PropTypes.object.isRequired
-};
-
-export default PermissionDialog;
