@@ -1,58 +1,45 @@
 import _ from 'lodash';
-import React, { Component } from 'react';
-import { ButtonToolbar } from 'react-bootstrap';
-import { Table, TableCell, TableAction, TableIconCell, TableBody, TableTextCell, TableHeader, TableColumn, TableRow } from '../Dashboard';
+import React, { PropTypes, Component } from 'react';
 
-class PermissionsTable extends Component {
+import PermissionRow from './PermissionRow';
+import { Table, TableBody, TableHeader, TableColumn } from 'auth0-extension-ui';
+
+export default class PermissionTable extends Component {
+  static propTypes = {
+    loading: PropTypes.bool,
+    permissions: PropTypes.object.isRequired,
+    applications: PropTypes.object.isRequired,
+    renderActions: PropTypes.func.isRequired
+  };
+
   shouldComponentUpdate(nextProps) {
     return nextProps.permissions !== this.props.permissions || nextProps.applications !== this.props.applications;
   }
 
   render() {
-    const applications = this.props.applications.toJS();
-    const permissions = this.props.permissions.toJS().map(perm => {
-      const application = _.find(applications, { client_id: perm.client_id });
-      perm.client_name = application && application.name || perm.client_id || 'N/A';
-      return perm;
-    });
+    const permissions = this.props.permissions.toJS();
+    const applications = this.props.applications.toJS().records;
 
-    return <Table>
-      <TableHeader>
-        <TableColumn width="3%"></TableColumn>
-        <TableColumn width="18%">Name</TableColumn>
-        <TableColumn width="33%">Description</TableColumn>
-        <TableColumn width="30%">Application</TableColumn>
-        <TableColumn width="15%"></TableColumn>
-      </TableHeader>
-      <TableBody>
-        {_.sortBy(permissions, 'client_name').map((permission, index) => {
-          return <TableRow key={index}>
-              <TableIconCell icon="384" />
-              <TableTextCell>{ permission.name || 'N/A' }</TableTextCell>
-              <TableTextCell>{ permission.description || 'N/A' }</TableTextCell>
-              <TableTextCell>{ permission.client_name }</TableTextCell>
-              <TableCell>
-                <ButtonToolbar style={{ marginBottom: '0px' }}>
-                  <TableAction id={`edit-${index}`} type="default" title="Edit Permission" icon="266"
-                    onClick={() => this.props.onEdit(permission)} disabled={this.props.loading || false} />
-                  <TableAction id={`delete-${index}`} type="success" title="Delete Permission" icon="263"
-                    onClick={() => this.props.onDelete(permission)} disabled={this.props.loading || false} />
-                </ButtonToolbar>
-              </TableCell>
-            </TableRow>;
-        })
-      }
-      </TableBody>
-    </Table>;
+    return (
+      <Table>
+        <TableHeader>
+          <TableColumn width="30%">Name</TableColumn>
+          <TableColumn width="26%">Application</TableColumn>
+          <TableColumn width="32%">Description</TableColumn>
+          <TableColumn />
+        </TableHeader>
+        <TableBody>
+          {_.sortBy(permissions, 'name').map((permission, index) => {
+            const application = applications.find(app => app.client_id === permission.applicationId);
+            return (
+              <PermissionRow
+                key={index} index={index} application={application} permission={permission}
+                renderActions={this.props.renderActions}
+              />
+            );
+          })}
+        </TableBody>
+      </Table>
+    );
   }
 }
-
-PermissionsTable.propTypes = {
-  onEdit: React.PropTypes.func.isRequired,
-  onDelete: React.PropTypes.func.isRequired,
-  applications: React.PropTypes.object.isRequired,
-  permissions: React.PropTypes.object.isRequired,
-  loading: React.PropTypes.bool.isRequired
-};
-
-export default PermissionsTable;
