@@ -1,20 +1,12 @@
 import React, { Component } from 'react';
 import { ButtonToolbar } from 'react-bootstrap';
 
-import { TableActionCell, Table, TableCell, TableRouteCell, TableBody, TableTextCell, TableHeader, TableColumn, TableRow } from '../Dashboard';
+import { TableActionCell, Table, TableCell, TableRouteCell, TableBody, TableTextCell, TableHeader, TableColumn, TableRow } from 'auth0-extension-ui';
 
 class UsersTable extends Component {
-  constructor() {
-    super();
-    this.renderActions = this.renderActions.bind(this);
-  }
 
   shouldComponentUpdate(nextProps) {
     return nextProps.users !== this.props.users;
-  }
-
-  renderActions(user, index) {
-    return this.props.renderActions(user, index);
   }
 
   render() {
@@ -27,27 +19,23 @@ class UsersTable extends Component {
           <TableColumn width="29%">Email</TableColumn>
           <TableColumn width="15%">Latest Login</TableColumn>
           <TableColumn width="10%">Logins</TableColumn>
-          <TableColumn width="17%">Connection</TableColumn>
-          <TableColumn width="15%" />
+          <TableColumn width="25%">Connection</TableColumn>
+          { this.props.renderActions ? <TableColumn width="8%" /> : null }
         </TableHeader>
         <TableBody>
-        {users.map((user, index) => {
-          return (
-              <TableRow key={index}>
-                <TableCell>
-                  <img className="img-circle" src={ user.picture } alt={ user.name || user.email || user.user_id } width="32" />
-                </TableCell>
-                <TableRouteCell route={`/users/${user.user_id}`}>{ user.name || user.email || user.user_id }</TableRouteCell>
-                <TableTextCell>{ user.email || 'N/A' }</TableTextCell>
-                <TableTextCell>{ user.last_login_relative }</TableTextCell>
-                <TableTextCell>{ user.logins_count }</TableTextCell>
-                <TableTextCell>{ user.identities[0].connection }</TableTextCell>
-                <TableActionCell>
-                  {this.renderActions(user, index)}
-                </TableActionCell>
-              </TableRow>
-            );
-        })}
+          {users.map((user, index) => (
+            <TableRow key={index}>
+              <TableCell>
+                <img className="img-circle" src={user.picture} alt={user.name || user.email || user.user_id} width="32" />
+              </TableCell>
+              <TableRouteCell route={`/users/${user.user_id}`}>{ user.name || user.email || user.user_id }</TableRouteCell>
+              <TableTextCell>{ user.email || 'N/A' }</TableTextCell>
+              <TableTextCell>{ user.last_login_relative }</TableTextCell>
+              <TableTextCell>{ user.logins_count }</TableTextCell>
+              <TableTextCell>{ user.identities[0].connection }</TableTextCell>
+              { this.props.renderActions ? <TableCell>{this.props.renderActions(user, index)}</TableCell> : null }
+            </TableRow>
+            ))}
         </TableBody>
       </Table>
     );
