@@ -8,46 +8,6 @@ export const FETCH_APPLICATIONS_PENDING = 'FETCH_APPLICATIONS_PENDING';
 export const FETCH_APPLICATIONS_REJECTED = 'FETCH_APPLICATIONS_REJECTED';
 export const FETCH_APPLICATIONS_FULFILLED = 'FETCH_APPLICATIONS_FULFILLED';
 
-// Fetch single.
-export const FETCH_APPLICATION = 'FETCH_APPLICATION';
-export const FETCH_APPLICATION_PENDING = 'FETCH_APPLICATION_PENDING';
-export const FETCH_APPLICATION_REJECTED = 'FETCH_APPLICATION_REJECTED';
-export const FETCH_APPLICATION_FULFILLED = 'FETCH_APPLICATION_FULFILLED';
-
-/*
- * Application Groups
- */
-
-export const FETCH_APPLICATION_GROUPS = 'FETCH_APPLICATION_GROUPS';
-export const FETCH_APPLICATION_GROUPS_PENDING = 'FETCH_APPLICATION_GROUPS_PENDING';
-export const FETCH_APPLICATION_GROUPS_REJECTED = 'FETCH_APPLICATION_GROUPS_REJECTED';
-export const FETCH_APPLICATION_GROUPS_FULFILLED = 'FETCH_APPLICATION_GROUPS_FULFILLED';
-
-// Add.
-export const ADD_APPLICATION_GROUP = 'ADD_APPLICATION_GROUP';
-export const ADD_APPLICATION_GROUP_PENDING = 'ADD_APPLICATION_GROUP_PENDING';
-export const ADD_APPLICATION_GROUP_REJECTED = 'ADD_APPLICATION_GROUP_REJECTED';
-export const ADD_APPLICATION_GROUP_FULFILLED = 'ADD_APPLICATION_GROUP_FULFILLED';
-
-// Create or edit.
-export const CREATE_APPLICATION_GROUP = 'CREATE_APPLICATION_GROUP';
-export const EDIT_APPLICATION_GROUP = 'EDIT_APPLICATION_GROUP';
-export const CLEAR_APPLICATION_GROUP = 'CLEAR_APPLICATION_GROUP';
-
-// Save.
-export const SAVE_APPLICATION_GROUP = 'SAVE_APPLICATION_GROUP';
-export const SAVE_APPLICATION_GROUP_PENDING = 'SAVE_APPLICATION_GROUP_PENDING';
-export const SAVE_APPLICATION_GROUP_REJECTED = 'SAVE_APPLICATION_GROUP_REJECTED';
-export const SAVE_APPLICATION_GROUP_FULFILLED = 'SAVE_APPLICATION_GROUP_FULFILLED';
-
-// Delete.
-export const REQUEST_REMOVE_APPLICATION_GROUP = 'REQUEST_REMOVE_APPLICATION_GROUP';
-export const CANCEL_REMOVE_APPLICATION_GROUP = 'CANCEL_REMOVE_APPLICATION_GROUP';
-export const REMOVE_APPLICATION_GROUP = 'REMOVE_APPLICATION_GROUP';
-export const REMOVE_APPLICATION_GROUP_PENDING = 'REMOVE_APPLICATION_GROUP_PENDING';
-export const REMOVE_APPLICATION_GROUP_REJECTED = 'REMOVE_APPLICATION_GROUP_REJECTED';
-export const REMOVE_APPLICATION_GROUP_FULFILLED = 'REMOVE_APPLICATION_GROUP_FULFILLED';
-
 /*
  * Auth.
  */
@@ -77,6 +37,60 @@ export const FETCH_CONNECTIONS_REJECTED = 'FETCH_CONNECTIONS_REJECTED';
 export const FETCH_CONNECTIONS_FULFILLED = 'FETCH_CONNECTIONS_FULFILLED';
 
 /*
+ * Configuration.
+ */
+
+// Fetch.
+export const FETCH_RULE_STATUS = 'FETCH_RULE_STATUS';
+export const FETCH_RULE_STATUS_PENDING = 'FETCH_RULE_STATUS_PENDING';
+export const FETCH_RULE_STATUS_REJECTED = 'FETCH_RULE_STATUS_REJECTED';
+export const FETCH_RULE_STATUS_FULFILLED = 'FETCH_RULE_STATUS_FULFILLED';
+
+// Fetch.
+export const FETCH_CONFIGURATION = 'FETCH_CONFIGURATION';
+export const FETCH_CONFIGURATION_PENDING = 'FETCH_CONFIGURATION_PENDING';
+export const FETCH_CONFIGURATION_REJECTED = 'FETCH_CONFIGURATION_REJECTED';
+export const FETCH_CONFIGURATION_FULFILLED = 'FETCH_CONFIGURATION_FULFILLED';
+
+// Save.
+export const SAVE_CONFIGURATION = 'SAVE_CONFIGURATION';
+export const SAVE_CONFIGURATION_PENDING = 'SAVE_CONFIGURATION_PENDING';
+export const SAVE_CONFIGURATION_REJECTED = 'SAVE_CONFIGURATION_REJECTED';
+export const SAVE_CONFIGURATION_FULFILLED = 'SAVE_CONFIGURATION_FULFILLED';
+
+/*
+ * Configuration Resource Server.
+ */
+
+// Fetch.
+export const FETCH_CONFIGURATION_RESOURCESERVER = 'FETCH_CONFIGURATION_RESOURCESERVER';
+export const FETCH_CONFIGURATION_RESOURCESERVER_PENDING = 'FETCH_CONFIGURATION_RESOURCESERVER_PENDING';
+export const FETCH_CONFIGURATION_RESOURCESERVER_REJECTED = 'FETCH_CONFIGURATION_RESOURCESERVER_REJECTED';
+export const FETCH_CONFIGURATION_RESOURCESERVER_FULFILLED = 'FETCH_CONFIGURATION_RESOURCESERVER_FULFILLED';
+
+// Save.
+export const SAVE_CONFIGURATION_RESOURCESERVER = 'SAVE_CONFIGURATION_RESOURCESERVER';
+export const SAVE_CONFIGURATION_RESOURCESERVER_PENDING = 'SAVE_CONFIGURATION_RESOURCESERVER_PENDING';
+export const SAVE_CONFIGURATION_RESOURCESERVER_REJECTED = 'SAVE_CONFIGURATION_RESOURCESERVER_REJECTED';
+export const SAVE_CONFIGURATION_RESOURCESERVER_FULFILLED = 'SAVE_CONFIGURATION_RESOURCESERVER_FULFILLED';
+
+// Delete.
+export const REMOVE_CONFIGURATION_RESOURCESERVER = 'REMOVE_CONFIGURATION_RESOURCESERVER';
+export const REMOVE_CONFIGURATION_RESOURCESERVER_PENDING = 'REMOVE_CONFIGURATION_RESOURCESERVER_PENDING';
+export const REMOVE_CONFIGURATION_RESOURCESERVER_REJECTED = 'REMOVE_CONFIGURATION_RESOURCESERVER_REJECTED';
+export const REMOVE_CONFIGURATION_RESOURCESERVER_FULFILLED = 'REMOVE_CONFIGURATION_RESOURCESERVER_FULFILLED';
+
+/*
+ * Configuration Explorer.
+ */
+
+ // Fetch.
+ export const FETCH_CONFIGURATION_EXPLORER = 'FETCH_CONFIGURATION_EXPLORER';
+ export const FETCH_CONFIGURATION_EXPLORER_PENDING = 'FETCH_CONFIGURATION_EXPLORER_PENDING';
+ export const FETCH_CONFIGURATION_EXPLORER_REJECTED = 'FETCH_CONFIGURATION_EXPLORER_REJECTED';
+ export const FETCH_CONFIGURATION_EXPLORER_FULFILLED = 'FETCH_CONFIGURATION_EXPLORER_FULFILLED';
+
+/*
  * Groups.
  */
 
@@ -85,6 +99,8 @@ export const FETCH_GROUPS = 'FETCH_GROUPS';
 export const FETCH_GROUPS_PENDING = 'FETCH_GROUPS_PENDING';
 export const FETCH_GROUPS_REJECTED = 'FETCH_GROUPS_REJECTED';
 export const FETCH_GROUPS_FULFILLED = 'FETCH_GROUPS_FULFILLED';
+export const CLOSE_UPDATE_GROUP = 'CLOSE_UPDATE_GROUP';
+export const CLOSE_DELETE_GROUP = 'CLOSE_DELETE_GROUP';
 
 // Fetch single.
 export const FETCH_GROUP = 'FETCH_GROUP';
@@ -98,9 +114,15 @@ export const SAVE_GROUP_PENDING = 'SAVE_GROUP_PENDING';
 export const SAVE_GROUP_REJECTED = 'SAVE_GROUP_REJECTED';
 export const SAVE_GROUP_FULFILLED = 'SAVE_GROUP_FULFILLED';
 
+export const UPDATE_GROUP = 'UPDATE_GROUP';
+export const UPDATE_GROUP_PENDING = 'UPDATE_GROUP_PENDING';
+export const UPDATE_GROUP_REJECTED = 'UPDATE_GROUP_REJECTED';
+export const UPDATE_GROUP_FULFILLED = 'UPDATE_GROUP_FULFILLED';
+
 // Create or edit.
 export const CREATE_GROUP = 'CREATE_GROUP';
 export const EDIT_GROUP = 'EDIT_GROUP';
+export const EDIT_GROUP_USERS = 'EDIT_GROUP_USERS';
 export const CLEAR_GROUP = 'CLEAR_GROUP';
 
 // Delete.
@@ -210,32 +232,57 @@ export const LOAD_GROUP_PICKER_REJECTED = 'LOAD_GROUP_PICKER_REJECTED';
 export const LOAD_GROUP_PICKER_FULFILLED = 'LOAD_GROUP_PICKER_FULFILLED';
 
 /*
- * Logs.
+ * Roles.
  */
-
-export const CLEAR_LOG = 'CLEAR_LOG';
 
 // Fetch all.
-export const FETCH_LOGS = 'FETCH_LOGS';
-export const FETCH_LOGS_PENDING = 'FETCH_LOGS_PENDING';
-export const FETCH_LOGS_REJECTED = 'FETCH_LOGS_REJECTED';
-export const FETCH_LOGS_FULFILLED = 'FETCH_LOGS_FULFILLED';
+export const FETCH_ROLES = 'FETCH_ROLES';
+export const FETCH_ROLES_PENDING = 'FETCH_ROLES_PENDING';
+export const FETCH_ROLES_REJECTED = 'FETCH_ROLES_REJECTED';
+export const FETCH_ROLES_FULFILLED = 'FETCH_ROLES_FULFILLED';
 
 // Fetch single.
-export const FETCH_LOG = 'FETCH_LOG';
-export const FETCH_LOG_PENDING = 'FETCH_LOG_PENDING';
-export const FETCH_LOG_REJECTED = 'FETCH_LOG_REJECTED';
-export const FETCH_LOG_FULFILLED = 'FETCH_LOG_FULFILLED';
+export const FETCH_ROLE = 'FETCH_ROLE';
+export const FETCH_ROLE_PENDING = 'FETCH_ROLE_PENDING';
+export const FETCH_ROLE_REJECTED = 'FETCH_ROLE_REJECTED';
+export const FETCH_ROLE_FULFILLED = 'FETCH_ROLE_FULFILLED';
+
+// Save.
+export const SAVE_ROLE = 'SAVE_ROLE';
+export const SAVE_ROLE_PENDING = 'SAVE_ROLE_PENDING';
+export const SAVE_ROLE_REJECTED = 'SAVE_ROLE_REJECTED';
+export const SAVE_ROLE_FULFILLED = 'SAVE_ROLE_FULFILLED';
+
+// Create or edit.
+export const CREATE_ROLE = 'CREATE_ROLE';
+export const EDIT_ROLE = 'EDIT_ROLE';
+export const CLEAR_ROLE = 'CLEAR_ROLE';
+
+export const ROLE_APPLICATION_SELECTED = 'ROLE_APPLICATION_SELECTED';
+
+// Delete.
+export const REQUEST_DELETE_ROLE = 'REQUEST_DELETE_ROLE';
+export const CANCEL_DELETE_ROLE = 'CANCEL_DELETE_ROLE';
+export const DELETE_ROLE = 'DELETE_ROLE';
+export const DELETE_ROLE_PENDING = 'DELETE_ROLE_PENDING';
+export const DELETE_ROLE_REJECTED = 'DELETE_ROLE_REJECTED';
+export const DELETE_ROLE_FULFILLED = 'DELETE_ROLE_FULFILLED';
 
 /*
- * Permissions
+ * Permissions.
  */
 
-// Fetch.
+// Fetch all.
 export const FETCH_PERMISSIONS = 'FETCH_PERMISSIONS';
 export const FETCH_PERMISSIONS_PENDING = 'FETCH_PERMISSIONS_PENDING';
 export const FETCH_PERMISSIONS_REJECTED = 'FETCH_PERMISSIONS_REJECTED';
 export const FETCH_PERMISSIONS_FULFILLED = 'FETCH_PERMISSIONS_FULFILLED';
+
+// Fetch single.
+export const FETCH_PERMISSION = 'FETCH_PERMISSION';
+export const FETCH_PERMISSION_PENDING = 'FETCH_PERMISSION_PENDING';
+export const FETCH_PERMISSION_REJECTED = 'FETCH_PERMISSION_REJECTED';
+export const FETCH_PERMISSION_FULFILLED = 'FETCH_PERMISSION_FULFILLED';
 
 // Save.
 export const SAVE_PERMISSION = 'SAVE_PERMISSION';
@@ -243,9 +290,9 @@ export const SAVE_PERMISSION_PENDING = 'SAVE_PERMISSION_PENDING';
 export const SAVE_PERMISSION_REJECTED = 'SAVE_PERMISSION_REJECTED';
 export const SAVE_PERMISSION_FULFILLED = 'SAVE_PERMISSION_FULFILLED';
 
-// Create or Edit.
+// Create or edit.
 export const CREATE_PERMISSION = 'CREATE_PERMISSION';
-export const EDIT_PERMISSION = 'SAVE_PERMISSION';
+export const EDIT_PERMISSION = 'EDIT_PERMISSION';
 export const CLEAR_PERMISSION = 'CLEAR_PERMISSION';
 
 // Delete.
@@ -257,35 +304,6 @@ export const DELETE_PERMISSION_REJECTED = 'DELETE_PERMISSION_REJECTED';
 export const DELETE_PERMISSION_FULFILLED = 'DELETE_PERMISSION_FULFILLED';
 
 /*
- * Roles.
- */
-
-// Fetch.
-export const FETCH_ROLES = 'FETCH_ROLES';
-export const FETCH_ROLES_PENDING = 'FETCH_ROLES_PENDING';
-export const FETCH_ROLES_REJECTED = 'FETCH_ROLES_REJECTED';
-export const FETCH_ROLES_FULFILLED = 'FETCH_ROLES_FULFILLED';
-
-// Save.
-export const SAVE_ROLE = 'SAVE_ROLE';
-export const SAVE_ROLE_PENDING = 'SAVE_ROLE_PENDING';
-export const SAVE_ROLE_REJECTED = 'SAVE_ROLE_REJECTED';
-export const SAVE_ROLE_FULFILLED = 'SAVE_ROLE_FULFILLED';
-
-// Create or Edit.
-export const CREATE_ROLE = 'CREATE_ROLE';
-export const EDIT_ROLE = 'SAVE_ROLE';
-export const CLEAR_ROLE = 'CLEAR_ROLE';
-
-// Delete.
-export const REQUEST_DELETE_ROLE = 'REQUEST_DELETE_ROLE';
-export const CANCEL_DELETE_ROLE = 'CANCEL_DELETE_ROLE';
-export const DELETE_ROLE = 'DELETE_ROLE';
-export const DELETE_ROLE_PENDING = 'DELETE_ROLE_PENDING';
-export const DELETE_ROLE_REJECTED = 'DELETE_ROLE_REJECTED';
-export const DELETE_ROLE_FULFILLED = 'DELETE_ROLE_FULFILLED';
-
-/*
  * Users.
  */
 
@@ -295,6 +313,9 @@ export const FETCH_USERS_PENDING = 'FETCH_USERS_PENDING';
 export const FETCH_USERS_REJECTED = 'FETCH_USERS_REJECTED';
 export const FETCH_USERS_FULFILLED = 'FETCH_USERS_FULFILLED';
 
+// Reset Fetched Users.
+export const RESET_FETCH_USERS = 'RESET_FETCH_USERS';
+
 // Fetch single.
 export const FETCH_USER = 'FETCH_USER';
 export const FETCH_USER_PENDING = 'FETCH_USER_PENDING';
@@ -302,10 +323,10 @@ export const FETCH_USER_REJECTED = 'FETCH_USER_REJECTED';
 export const FETCH_USER_FULFILLED = 'FETCH_USER_FULFILLED';
 
 // Fetch user authorization.
-export const FETCH_USER_AUTHORIZATION = 'FETCH_USER_AUTHORIZATION';
-export const FETCH_USER_AUTHORIZATION_PENDING = 'FETCH_USER_AUTHORIZATION_PENDING';
-export const FETCH_USER_AUTHORIZATION_REJECTED = 'FETCH_USER_AUTHORIZATION_REJECTED';
-export const FETCH_USER_AUTHORIZATION_FULFILLED = 'FETCH_USER_AUTHORIZATION_FULFILLED';
+export const FETCH_USER_NESTED_GROUPS = 'FETCH_USER_NESTED_GROUPS';
+export const FETCH_USER_NESTED_GROUPS_PENDING = 'FETCH_USER_NESTED_GROUPS_PENDING';
+export const FETCH_USER_NESTED_GROUPS_REJECTED = 'FETCH_USER_NESTED_GROUPS_REJECTED';
+export const FETCH_USER_NESTED_GROUPS_FULFILLED = 'FETCH_USER_NESTED_GROUPS_FULFILLED';
 
 // Fetch logs.
 export const FETCH_USER_LOGS = 'FETCH_USER_LOGS';
@@ -368,3 +389,67 @@ export const SEARCH_USER_PICKER_FULFILLED = 'SEARCH_USER_PICKER_FULFILLED';
 // Select.
 export const SELECT_USER = 'SELECT_USER';
 export const UNSELECT_USER = 'UNSELECT_USER';
+
+// Configuration
+export const FETCH_CONFIGURATION_EXPORT = 'FETCH_CONFIGURATION_EXPORT';
+export const FETCH_CONFIGURATION_EXPORT_PENDING = 'FETCH_CONFIGURATION_EXPORT_PENDING';
+export const FETCH_CONFIGURATION_EXPORT_REJECTED = 'FETCH_CONFIGURATION_EXPORT_REJECTED';
+export const FETCH_CONFIGURATION_EXPORT_FULFILLED = 'FETCH_CONFIGURATION_EXPORT_FULFILLED';
+
+export const FETCH_CONFIGURATION_IMPORT = 'FETCH_CONFIGURATION_IMPORT';
+export const FETCH_CONFIGURATION_IMPORT_PENDING = 'FETCH_CONFIGURATION_IMPORT_PENDING';
+export const FETCH_CONFIGURATION_IMPORT_REJECTED = 'FETCH_CONFIGURATION_IMPORT_REJECTED';
+export const FETCH_CONFIGURATION_IMPORT_FULFILLED = 'FETCH_CONFIGURATION_IMPORT_FULFILLED';
+
+export const FETCH_CONFIGURATION_ADD_ERROR = 'FETCH_CONFIGURATION_ADD_ERROR';
+export const FETCH_CONFIGURATION_CLOSE_ERROR = 'FETCH_CONFIGURATION_CLOSE_ERROR';
+export const OPEN_CONFIGURATION_PREVIEW = 'OPEN_CONFIGURATION_PREVIEW';
+export const CLOSE_CONFIGURATION_PREVIEW = 'CLOSE_CONFIGURATION_PREVIEW';
+export const SET_CONFIGURATION_TAB = 'SET_CONFIGURATION_TAB';
+
+
+export const ADD_ROLES_OPEN = 'ADD_ROLES_OPEN';
+export const ADD_ROLES_CLOSE = 'ADD_ROLES_CLOSE';
+
+export const SAVE_USER_ROLES = 'SAVE_USER_ROLES';
+export const SAVE_USER_ROLES_PENDING = 'SAVE_USER_ROLES_PENDING';
+export const SAVE_USER_ROLES_REJECTED = 'SAVE_USER_ROLES_REJECTED';
+export const SAVE_USER_ROLES_FULFILLED = 'SAVE_USER_ROLES_FULFILLED';
+
+export const FETCH_USER_ROLES = 'FETCH_USER_ROLES';
+export const FETCH_USER_ROLES_PENDING = 'FETCH_USER_ROLES_PENDING';
+export const FETCH_USER_ROLES_REJECTED = 'FETCH_USER_ROLES_REJECTED';
+export const FETCH_USER_ROLES_FULFILLED = 'FETCH_USER_ROLES_FULFILLED';
+
+export const FETCH_USER_ALL_ROLES = 'FETCH_USER_ALL_ROLES';
+export const FETCH_USER_ALL_ROLES_PENDING = 'FETCH_USER_ALL_ROLES_PENDING';
+export const FETCH_USER_ALL_ROLES_REJECTED = 'FETCH_USER_ALL_ROLES_REJECTED';
+export const FETCH_USER_ALL_ROLES_FULFILLED = 'FETCH_USER_ALL_ROLES_FULFILLED';
+
+export const GROUP_ADD_ROLES_OPEN = 'GROUP_ADD_ROLES_OPEN';
+export const GROUP_ADD_ROLES_CLOSE = 'GROUP_ADD_ROLES_CLOSE';
+
+export const SAVE_GROUP_ROLES = 'SAVE_GROUP_ROLES';
+export const SAVE_GROUP_ROLES_PENDING = 'SAVE_GROUP_ROLES_PENDING';
+export const SAVE_GROUP_ROLES_REJECTED = 'SAVE_GROUP_ROLES_REJECTED';
+export const SAVE_GROUP_ROLES_FULFILLED = 'SAVE_GROUP_ROLES_FULFILLED';
+
+export const FETCH_GROUP_ROLES = 'FETCH_GROUP_ROLES';
+export const FETCH_GROUP_ROLES_PENDING = 'FETCH_GROUP_ROLES_PENDING';
+export const FETCH_GROUP_ROLES_REJECTED = 'FETCH_GROUP_ROLES_REJECTED';
+export const FETCH_GROUP_ROLES_FULFILLED = 'FETCH_GROUP_ROLES_FULFILLED';
+
+
+export const REQUEST_DELETE_GROUP_ROLE = 'REQUEST_DELETE_GROUP_ROLE';
+export const CANCEL_DELETE_GROUP_ROLE = 'CANCEL_DELETE_GROUP_ROLE';
+export const DELETE_GROUP_ROLE = 'DELETE_GROUP_ROLE';
+export const DELETE_GROUP_ROLE_PENDING = 'DELETE_GROUP_ROLE_PENDING';
+export const DELETE_GROUP_ROLE_REJECTED = 'DELETE_GROUP_ROLE_REJECTED';
+export const DELETE_GROUP_ROLE_FULFILLED = 'DELETE_GROUP_ROLE_FULFILLED';
+
+export const REQUEST_DELETE_USER_ROLE = 'REQUEST_DELETE_USER_ROLE';
+export const CANCEL_DELETE_USER_ROLE = 'CANCEL_DELETE_USER_ROLE';
+export const DELETE_USER_ROLE = 'DELETE_USER_ROLE';
+export const DELETE_USER_ROLE_PENDING = 'DELETE_USER_ROLE_PENDING';
+export const DELETE_USER_ROLE_REJECTED = 'DELETE_USER_ROLE_REJECTED';
+export const DELETE_USER_ROLE_FULFILLED = 'DELETE_USER_ROLE_FULFILLED';
