@@ -1,13 +1,11 @@
 export App from './App.jsx';
-export ApplicationContainer from './Applications/ApplicationContainer';
-export ApplicationsContainer from './Applications/ApplicationsContainer';
-export LogsContainer from './Logs/LogsContainer';
-export LoginContainer from './Login/LoginContainer';
-export GroupContainer from './Groups/GroupContainer';
-export GroupsContainer from './Groups/GroupsContainer';
-export PermissionContainer from './Permissions/PermissionContainer';
-export PermissionsContainer from './Permissions/PermissionsContainer';
-export RoleContainer from './Roles/RoleContainer';
-export RolesContainer from './Roles/RolesContainer';
-export UserContainer from './Users/UserContainer';
-export UsersContainer from './Users/UsersContainer';
+export API from './API';
+export ConfigurationRule from './ConfigurationRule';
+export Group from './Group';
+export Groups from './Groups';
+export Login from './Login';
+export Permissions from './Permissions';
+export RequireAuthentication from './RequireAuthentication';
+export Roles from './Roles';
+export User from './User';
+export Users from './Users';
