@@ -4,8 +4,7 @@ import promiseMiddleware from 'redux-promise-middleware';
 import { compose, createStore, applyMiddleware } from 'redux';
 
 import rootReducer from '../reducers';
-import normalizeErrorMiddleware from '../middlewares/normalizeErrorMiddleware';
-import promiseSuccessMiddleware from '../middlewares/promiseSuccessMiddleware';
+import { promiseSuccessMiddleware, normalizeErrorMiddleware } from 'auth0-extension-ui-redux';
 import DevTools from '../containers/DevTools';
 
 
