@@ -1,4 +1,5 @@
 const winston = require('winston');
+
 winston.emitErrs = true;
 
 const logger = new winston.Logger({
@@ -15,8 +16,3 @@ const logger = new winston.Logger({
 });
 
 module.exports = logger;
-module.exports.stream = {
-  write: (message) => {
-    logger.info(message.replace(/\n$/, ''));
-  }
-};
